change /bin/bash link to be wrapped with if to check if file exsists

+++ configuration.nix
@@ -86,7 +86,9 @@
   system.activationScripts.binbash = {
     deps = ["binsh"];
     text = ''
-      ln -s /bin/sh /bin/bash
+      if [ ! -e "/bin/bash" ]; then
+      	ln -s /bin/sh /bin/bash
+      fi
     '';
   };
